refactor: separate agent lifecycle helpers
