+++ src/agent/loop_run/lifecycle.rs
@@ -1,0 +1,132 @@
+use super::*;
+
+impl Agent {
+    pub(super) fn current_plan_contents(&self) -> Result<Option<String>, String> {
+        let Some(path) = &self.session.mode_state.active_plan_path else {
+            return Ok(None);
+        };
+        if !path.exists() {
+            return Ok(None);
+        }
+        let contents = std::fs::read_to_string(path)
+            .map_err(|err| format!("failed to read plan {}: {err}", path.display()))?;
+        Ok(Some(contents))
+    }
+
+    pub(super) fn persist_session(&mut self) -> Result<(), String> {
+        self.session.active_root = if self.work_root == self.config.cwd {
+            None
+        } else {
+            Some(self.work_root.clone())
+        };
+        self.session_store.save(&self.session)
+    }
+
+    pub(super) fn maybe_compact_session(&mut self, keep_tail: usize) -> bool {
+        if !should_compact(
+            &self.session.messages,
+            self.config.context_budget,
+            keep_tail,
+        ) {
+            return false;
+        }
+
+        if let Some(sidecar) = self.models.sidecar.clone() {
+            compact_messages_with_strategy(&mut self.session.messages, keep_tail, |head| {
+                self.client.summarize_conversation(&sidecar, head)
+            })
+            .unwrap_or_else(|_| compact_messages(&mut self.session.messages, keep_tail))
+        } else {
+            compact_messages(&mut self.session.messages, keep_tail)
+        }
+    }
+
+    pub(super) fn ensure_plan_file(&self, plan_path: &Path) -> Result<(), String> {
+        if plan_path.exists() {
+            return Ok(());
+        }
+        if let Some(parent) = plan_path.parent() {
+            std::fs::create_dir_all(parent)
+                .map_err(|err| format!("failed to create {}: {err}", parent.display()))?;
+        }
+        std::fs::write(
+            plan_path,
+            "# Plan\n\n## Goal\n- \n\n## Findings\n- \n\n## Steps\n1. \n",
+        )
+        .map_err(|err| format!("failed to create plan file {}: {err}", plan_path.display()))
+    }
+
+    pub(super) fn maybe_update_work_root(
+        &mut self,
+        name: &str,
+        arguments: &serde_json::Value,
+        result: &str,
+    ) {
+        let Some(new_root) = prompting::detect_scaffold_root(name, arguments, result) else {
+            return;
+        };
+        if new_root == self.work_root || !new_root.is_dir() {
+            return;
+        }
+
+        self.apply_scaffold_root(new_root);
+    }
+
+    pub(super) fn disable_native_tools_for_session(&mut self) {
+        self.native_tools_enabled = false;
+        self.session.native_tools_disabled = true;
+        self.push_system_note(
+            "[Runtime Notice] Native tool calling is disabled for this session because the model/runtime parser rejected a tool-call response. Use only <anvil_tool_call>{\"name\":\"Tool\",\"arguments\":{...}}</anvil_tool_call> blocks with valid JSON arguments."
+                .to_string(),
+        );
+    }
+
+    fn apply_scaffold_root(&mut self, new_root: PathBuf) {
+        self.work_root = new_root.clone();
+        self.session.active_root = Some(new_root.clone());
+        self.reset_after_scaffold();
+        self.push_system_note(format!(
+            "[Workspace Root Updated] Continue work inside {} and use relative paths from there.",
+            new_root.display()
+        ));
+    }
+
+    fn reset_after_scaffold(&mut self) {
+        self.session.messages =
+            prompting::reset_messages_after_scaffold(&self.work_root, &self.session.messages);
+    }
+}
+
+pub(super) fn should_compact(
+    messages: &[ConversationMessage],
+    context_budget: usize,
+    keep_tail: usize,
+) -> bool {
+    messages.len() > keep_tail + 4 || approximate_token_count(messages) > context_budget
+}
+
+pub(super) fn is_native_tool_parser_failure(error: &str) -> bool {
+    let lower = error.to_ascii_lowercase();
+    lower.contains("native tool parser failed")
+        || lower.contains("unexpected end element")
+        || lower.contains("unexpected eof")
+}
+
+pub(super) fn plan_is_substantive(contents: &str) -> bool {
+    let meaningful_lines = contents
+        .lines()
+        .map(str::trim)
+        .filter(|line| !line.is_empty())
+        .filter(|line| !matches!(*line, "# Plan" | "## Goal" | "## Findings" | "## Steps"))
+        .filter(|line| *line != "-" && *line != "1.")
+        .count();
+    meaningful_lines >= 2
+}
+
+pub(super) fn format_tool_error(err: &str) -> String {
+    if err.starts_with("Error:") {
+        err.to_string()
+    } else {
+        format!("Error: {err}")
+    }
+}
